Move more logic to `mesh_pipeline_layout.rs`
